hackathon-ai/requirements.txt
python-dotenv
pydantic
pydantic-settings
openai
pandas
openpyxl
google-generativeai
groq
pypdf
chromadb
sentence-transformers
streamlit
